feat: :+1:  move flow classes to flows directory
